Move rgba_to_int to utils.py
